+++ packages/components/src/components/Icon/index.ts
@@ -1,5 +1,7 @@
 export * from "./view";
 
+export { iconCategories, type IconCategory } from "@mittwald/flow-icons";
+
 export { IconSetProvider } from "./components/IconSetProvider";
 export { type IconProps, Icon } from "./Icon";
 export { default } from "./Icon";