+++ packages/components/src/components/Message/Message.tsx
@@ -1,7 +1,10 @@
-import type { CSSProperties, PropsWithChildren } from "react";
+import { type CSSProperties, type PropsWithChildren } from "react";
 import clsx from "clsx";
 import styles from "./Message.module.scss";
-import type { PropsWithClassName } from "@/lib/types/props";
+import type {
+  PropsWithClassName,
+  PropsWithElementType,
+} from "@/lib/types/props";
 import {
   dynamic,
   type PropsContext,
@@ -13,9 +16,16 @@ import {
   type FlowComponentProps,
 } from "@/lib/componentFactory/flowComponent";
 import ClearPropsContext from "@/lib/propsContext/components/ClearPropsContext";
+import locales from "./locales/*.locale.json";
+import { useLocalizedStringFormatter } from "@/components/TranslationProvider";
+import * as Aria from "react-aria-components";
 
 export interface MessageProps
-  extends PropsWithChildren, PropsWithClassName, FlowComponentProps {
+  extends
+    PropsWithChildren,
+    PropsWithClassName,
+    FlowComponentProps,
+    PropsWithElementType<"article" | "li"> {
   /** Determines the color and orientation of the message. @default "responder" */
   type?: "responder" | "sender";
   color?: string;
@@ -23,7 +33,15 @@ export interface MessageProps
 
 /** @flr-generate all */
 export const Message = flowComponent("Message", (props) => {
-  const { type = "responder", children, className, color } = props;
+  const {
+    type = "responder",
+    children,
+    className,
+    color,
+    elementType = "article",
+  } = props;
+
+  const Element = elementType;
 
   const rootClassName = clsx(styles.message, styles[type], className);
 
@@ -31,14 +49,16 @@ export const Message = flowComponent("Message", (props) => {
     ? ({ "--message-background": color } as CSSProperties)
     : undefined;
 
+  const formatter = useLocalizedStringFormatter(locales, "Message");
+
   const propsContext: PropsContext = {
     Content: {
       className: styles.content,
       children: dynamic((props) => {
         return (
           <>
-            <div className={styles.tipBorder} />
-            <div className={styles.tip} />
+            <div className={styles.tipBorder} aria-hidden />
+            <div className={styles.tip} aria-hidden />
             {props.children}
           </>
         );
@@ -59,13 +79,23 @@ export const Message = flowComponent("Message", (props) => {
           variant: "plain",
           color: "secondary",
           children: <IconContextMenu />,
+          "aria-label": formatter.format("options"),
         },
       },
       Text: { className: styles.date },
       Align: {
         wrapWith: <ClearPropsContext />,
         className: styles.user,
+        Avatar: { Initials: { "aria-hidden": true } },
       },
+      children: dynamic((props) => (
+        <>
+          <Aria.VisuallyHidden>
+            {formatter.format("headerLabel")}
+          </Aria.VisuallyHidden>
+          {props.children}
+        </>
+      )),
     },
 
     Button: {
@@ -83,9 +113,9 @@ export const Message = flowComponent("Message", (props) => {
 
   return (
     <PropsContextProvider props={propsContext}>
-      <article className={rootClassName} style={style}>
+      <Element className={rootClassName} style={style}>
         {children}
-      </article>
+      </Element>
     </PropsContextProvider>
   );
 });